// octelium/cluster/portal/portal/web/package/src/apis/metav1/metav1.ts
export interface ListResponseMeta {
  /**
   * Page is the page number. Starts with zero.
   *
   * @generated from protobuf field: uint32 page = 1;
   */
  page: number;
  /**
   * ItemsPerPage is number of items per page.
   *
   * @generated from protobuf field: uint32 itemsPerPage = 2;
   */
  itemsPerPage: number;
  /**
   * TotalCount is the total count of items that can be obtained.
   *
   * @generated from protobuf field: uint32 totalCount = 3;
   */
  totalCount: number;
  /**
   * HasMore shows whether there a next page is available
   *
   * @generated from protobuf field: bool hasMore = 4;
   */
  hasMore: boolean;
}
// ...
class ListResponseMeta$Type extends MessageType<ListResponseMeta> {
  constructor() {
    super('octelium.api.main.meta.v1.ListResponseMeta', [
      { no: 1, name: 'page', kind: 'scalar', T: 13 /*ScalarType.UINT32*/ },
      { no: 2, name: 'itemsPerPage', kind: 'scalar', T: 13 /*ScalarType.UINT32*/ },
      { no: 3, name: 'totalCount', kind: 'scalar', T: 13 /*ScalarType.UINT32*/ },
      { no: 4, name: 'hasMore', kind: 'scalar', T: 8 /*ScalarType.BOOL*/ },
    ]);
  }
  create(value?: PartialMessage<ListResponseMeta>): ListResponseMeta {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.page = 0;
    message.itemsPerPage = 0;
    message.totalCount = 0;
    message.hasMore = false;
    if (value !== undefined) reflectionMergePartial<ListResponseMeta>(this, message, value);
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: ListResponseMeta
  ): ListResponseMeta {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* uint32 page */ 1:
          message.page = reader.uint32();
          break;
        case /* uint32 itemsPerPage */ 2:
          message.itemsPerPage = reader.uint32();
          break;
        case /* uint32 totalCount */ 3:
          message.totalCount = reader.uint32();
          break;
        case /* bool hasMore */ 4:
          message.hasMore = reader.bool();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: ListResponseMeta,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* uint32 page = 1; */
    if (message.page !== 0) writer.tag(1, WireType.Varint).uint32(message.page);
    /* uint32 itemsPerPage = 2; */
    if (message.itemsPerPage !== 0) writer.tag(2, WireType.Varint).uint32(message.itemsPerPage);
    /* uint32 totalCount = 3; */
    if (message.totalCount !== 0) writer.tag(3, WireType.Varint).uint32(message.totalCount);
    /* bool hasMore = 4; */
    if (message.hasMore !== false) writer.tag(4, WireType.Varint).bool(message.hasMore);
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.meta.v1.ListResponseMeta
 */
export const ListResponseMeta = new ListResponseMeta$Type();